refactor: create dtos folder in database module folder and move type definitons in
